Refactoring in progress 1/2
