+++ Commands/utility/resetnick.js
@@ -1,0 +1,8 @@
+module.exports = {
+    name: 'resetnick',
+    args: false,
+    execute(message, args){
+        const nick = message.author.tag;
+        message.member.setNickname(nick);
+    }
+}
