+++ token.json
@@ -1,0 +1,3 @@
+{
+    "BOT_TOKEN": "ODQwNzQ3NzkwNjg1OTYyMjQw.YJctZQ.zZDKXp0Dsf85LLDsTF6wkw3W1QU"
+}
